src/search_and_capture_algo/requirements.txt
gurobipy>=12.0.0
numpy>=1.20.2
pycairo>=1.20.0
python-igraph>=0.9.1,<1.0.0
scipy>=1.6.2
texttable>=1.6.3
